afxdp/coordinator: relocate inline tests to coordinator/tests.rs
Mirrors the forwarding/, frame/, session_glue/, types/, and tx/
patterns. coordinator/mod.rs reaches 2,761 LOC (1,888 prod + 870 inline
tests + closing brace).

Mechanical split:
- inline #[cfg(test)] mod tests { ... } block (lines 1889-2760) moved to
  coordinator/tests.rs (870 lines, 19 tests)
- 4-space dedent applied uniformly to inner block
- mod.rs ends with: #[cfg(test)] mod tests; (auto-resolved within the
  coordinator/ directory module, no #[path] needed)

No production code touched. Use clauses kept verbatim at the top of
tests.rs (super::*, crate::test_zone_ids::*, crate::ClassOfServiceSnapshot
+ siblings) — module hierarchy is unchanged because tests still resolves
to super = coordinator via mod tests.

cargo build --release: 92 warnings baseline, no errors.
cargo test --release: 865 passed, 0 failed (baseline preserved).
