+++ userspace-dp/src/afxdp/coordinator/mod.rs
@@ -1887,875 +1887,4 @@ where
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_zone_ids::*;
-    use crate::{
-        ClassOfServiceSnapshot, CoSForwardingClassSnapshot, CoSSchedulerMapEntrySnapshot,
-        CoSSchedulerMapSnapshot,
-    };
-
-    #[test]
-    fn build_cos_owner_worker_by_queue_prefers_lowest_worker_with_tx_binding() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 1_000_000,
-                burst_bytes: 64 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![CoSQueueConfig {
-                    queue_id: 0,
-                    forwarding_class: "best-effort".into(),
-                    priority: 5,
-                    transmit_rate_bytes: 1_000_000,
-                    exact: false,
-                    surplus_weight: 1,
-                    buffer_bytes: 64 * 1024,
-                    dscp_rewrite: None,
-                }],
-            },
-        );
-        forwarding.egress.insert(
-            80,
-            EgressInterface {
-                bind_ifindex: 12,
-                vlan_id: 0,
-                mtu: 1500,
-                src_mac: [0; 6],
-                zone_id: TEST_WAN_ZONE_ID,
-                redundancy_group: 0,
-                primary_v4: None,
-                primary_v6: None,
-            },
-        );
-        let worker_binding_ifindexes = BTreeMap::from([
-            (2, std::collections::BTreeSet::from([12])),
-            (7, std::collections::BTreeSet::from([12, 13])),
-        ]);
-
-        let owner_by_queue = build_cos_owner_worker_by_queue_from_binding_ifindexes(
-            &forwarding,
-            &worker_binding_ifindexes,
-        );
-
-        assert_eq!(owner_by_queue.get(&(80, 0)), Some(&2));
-    }
-
-    #[test]
-    fn build_cos_owner_worker_by_queue_spreads_queues_across_eligible_workers() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 1_000_000,
-                burst_bytes: 64 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![
-                    CoSQueueConfig {
-                        queue_id: 0,
-                        forwarding_class: "best-effort".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 1_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 64 * 1024,
-                        dscp_rewrite: None,
-                    },
-                    CoSQueueConfig {
-                        queue_id: 1,
-                        forwarding_class: "af11".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 1_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 64 * 1024,
-                        dscp_rewrite: None,
-                    },
-                    CoSQueueConfig {
-                        queue_id: 2,
-                        forwarding_class: "af12".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 1_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 64 * 1024,
-                        dscp_rewrite: None,
-                    },
-                ],
-            },
-        );
-        forwarding.egress.insert(
-            80,
-            EgressInterface {
-                bind_ifindex: 12,
-                vlan_id: 0,
-                mtu: 1500,
-                src_mac: [0; 6],
-                zone_id: TEST_WAN_ZONE_ID,
-                redundancy_group: 0,
-                primary_v4: None,
-                primary_v6: None,
-            },
-        );
-        let worker_binding_ifindexes = BTreeMap::from([
-            (2, std::collections::BTreeSet::from([12])),
-            (7, std::collections::BTreeSet::from([12])),
-        ]);
-
-        let owner_by_queue = build_cos_owner_worker_by_queue_from_binding_ifindexes(
-            &forwarding,
-            &worker_binding_ifindexes,
-        );
-
-        assert_eq!(owner_by_queue.get(&(80, 0)), Some(&2));
-        assert_eq!(owner_by_queue.get(&(80, 1)), Some(&7));
-        assert_eq!(owner_by_queue.get(&(80, 2)), Some(&2));
-    }
-
-    #[test]
-    fn build_cos_owner_worker_by_queue_prefers_ready_workers_when_available() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 1_000_000,
-                burst_bytes: 64 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![
-                    CoSQueueConfig {
-                        queue_id: 0,
-                        forwarding_class: "best-effort".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 1_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 64 * 1024,
-                        dscp_rewrite: None,
-                    },
-                    CoSQueueConfig {
-                        queue_id: 4,
-                        forwarding_class: "iperf-a".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 1_000_000,
-                        exact: true,
-                        surplus_weight: 1,
-                        buffer_bytes: 64 * 1024,
-                        dscp_rewrite: None,
-                    },
-                ],
-            },
-        );
-        forwarding.egress.insert(
-            80,
-            EgressInterface {
-                bind_ifindex: 12,
-                vlan_id: 0,
-                mtu: 1500,
-                src_mac: [0; 6],
-                zone_id: TEST_WAN_ZONE_ID,
-                redundancy_group: 0,
-                primary_v4: None,
-                primary_v6: None,
-            },
-        );
-
-        let owner_by_queue = build_cos_owner_worker_by_queue_with_fallback_ifindexes(
-            &forwarding,
-            &BTreeMap::from([(7, std::collections::BTreeSet::from([12]))]),
-            &BTreeMap::from([
-                (2, std::collections::BTreeSet::from([12])),
-                (7, std::collections::BTreeSet::from([12])),
-            ]),
-        );
-
-        assert_eq!(owner_by_queue.get(&(80, 0)), Some(&7));
-        assert_eq!(owner_by_queue.get(&(80, 4)), Some(&7));
-    }
-
-    #[test]
-    fn build_cos_owner_worker_by_queue_falls_back_when_no_ready_workers_exist() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 1_000_000,
-                burst_bytes: 64 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![CoSQueueConfig {
-                    queue_id: 0,
-                    forwarding_class: "best-effort".into(),
-                    priority: 5,
-                    transmit_rate_bytes: 1_000_000,
-                    exact: false,
-                    surplus_weight: 1,
-                    buffer_bytes: 64 * 1024,
-                    dscp_rewrite: None,
-                }],
-            },
-        );
-        forwarding.egress.insert(
-            80,
-            EgressInterface {
-                bind_ifindex: 12,
-                vlan_id: 0,
-                mtu: 1500,
-                src_mac: [0; 6],
-                zone_id: TEST_WAN_ZONE_ID,
-                redundancy_group: 0,
-                primary_v4: None,
-                primary_v6: None,
-            },
-        );
-
-        let owner_by_queue = build_cos_owner_worker_by_queue_with_fallback_ifindexes(
-            &forwarding,
-            &BTreeMap::new(),
-            &BTreeMap::from([(2, std::collections::BTreeSet::from([12]))]),
-        );
-
-        assert_eq!(owner_by_queue.get(&(80, 0)), Some(&2));
-    }
-
-    #[test]
-    fn build_worker_binding_ifindexes_from_identities_groups_by_worker() {
-        let identities = BTreeMap::from([
-            (
-                10,
-                BindingIdentity {
-                    slot: 10,
-                    queue_id: 0,
-                    worker_id: 2,
-                    interface: "ge-0-0-2".into(),
-                    ifindex: 12,
-                },
-            ),
-            (
-                11,
-                BindingIdentity {
-                    slot: 11,
-                    queue_id: 1,
-                    worker_id: 2,
-                    interface: "ge-0-0-2".into(),
-                    ifindex: 12,
-                },
-            ),
-            (
-                20,
-                BindingIdentity {
-                    slot: 20,
-                    queue_id: 0,
-                    worker_id: 7,
-                    interface: "ge-0-0-3".into(),
-                    ifindex: 13,
-                },
-            ),
-        ]);
-
-        let worker_binding_ifindexes = build_worker_binding_ifindexes_from_identities(&identities);
-
-        assert_eq!(
-            worker_binding_ifindexes.get(&2),
-            Some(&std::collections::BTreeSet::from([12]))
-        );
-        assert_eq!(
-            worker_binding_ifindexes.get(&7),
-            Some(&std::collections::BTreeSet::from([13]))
-        );
-    }
-
-    #[test]
-    fn refresh_runtime_snapshot_rebuilds_cos_owner_worker_map_from_identities() {
-        let mut coordinator = Coordinator::new();
-        coordinator.workers.identities.insert(
-            1,
-            BindingIdentity {
-                slot: 1,
-                queue_id: 0,
-                worker_id: 2,
-                interface: "ge-0-0-2".into(),
-                ifindex: 12,
-            },
-        );
-        coordinator.workers.identities.insert(
-            2,
-            BindingIdentity {
-                slot: 2,
-                queue_id: 0,
-                worker_id: 7,
-                interface: "ge-0-0-3".into(),
-                ifindex: 13,
-            },
-        );
-
-        let mut snapshot = ConfigSnapshot::default();
-        snapshot.interfaces.push(InterfaceSnapshot {
-            name: "reth0.80".into(),
-            ifindex: 80,
-            parent_ifindex: 12,
-            hardware_addr: "02:00:00:00:00:80".into(),
-            cos_shaping_rate_bytes_per_sec: 1_000_000,
-            cos_scheduler_map: "wan-map".into(),
-            ..Default::default()
-        });
-        snapshot.class_of_service = Some(ClassOfServiceSnapshot {
-            forwarding_classes: vec![CoSForwardingClassSnapshot {
-                name: "best-effort".into(),
-                queue: 0,
-            }],
-            schedulers: vec![],
-            scheduler_maps: vec![CoSSchedulerMapSnapshot {
-                name: "wan-map".into(),
-                entries: vec![CoSSchedulerMapEntrySnapshot {
-                    forwarding_class: "best-effort".into(),
-                    scheduler: String::new(),
-                }],
-            }],
-            dscp_classifiers: vec![],
-            ieee8021_classifiers: vec![],
-            dscp_rewrite_rules: vec![],
-        });
-
-        coordinator.refresh_runtime_snapshot(&snapshot);
-
-        assert_eq!(
-            coordinator.cos_owner_worker_by_queue.get(&(80, 0)),
-            Some(&2)
-        );
-        let shared = coordinator.cos.owner_worker_by_queue.load();
-        assert_eq!(shared.get(&(80, 0)), Some(&2));
-    }
-
-    #[test]
-    fn build_shared_cos_root_leases_uses_active_workers_per_interface() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 100_000_000,
-                burst_bytes: 256 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![
-                    CoSQueueConfig {
-                        queue_id: 0,
-                        forwarding_class: "best-effort".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 50_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 128 * 1024,
-                        dscp_rewrite: None,
-                    },
-                    CoSQueueConfig {
-                        queue_id: 1,
-                        forwarding_class: "af11".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 50_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 128 * 1024,
-                        dscp_rewrite: None,
-                    },
-                ],
-            },
-        );
-        let active_shards_by_egress_ifindex = BTreeMap::from([(80, 2usize)]);
-
-        let leases = build_shared_cos_root_leases(&forwarding, &active_shards_by_egress_ifindex);
-        let lease = leases.get(&80).expect("shared root lease");
-
-        // The root lease budget must scale with active_shards: total
-        // grantable = lease_bytes * active_shards. That is the actual
-        // invariant this test pins. Drive it by reading active_shards
-        // from the fixture (so the assertion does not silently decouple
-        // from the setup) and drain the budget with fixed-size requests
-        // plus a tail remainder for whatever the budget does not cleanly
-        // divide by — lease_bytes is a function of shaping rate and
-        // COS_ROOT_LEASE_TARGET_US, both of which are tuning knobs, so
-        // an exact-divisibility assertion would make this test brittle
-        // against legitimate scheduler tuning.
-        let active_shards = *active_shards_by_egress_ifindex
-            .get(&80)
-            .expect("active shards configured for egress ifindex 80")
-            as u64;
-        let lease_bytes = lease.lease_bytes();
-        let expected_total = lease_bytes * active_shards;
-        let per_request = 2500u64;
-
-        let mut remaining = expected_total;
-        let mut total = 0u64;
-        while remaining > 0 {
-            let req = remaining.min(per_request);
-            let granted = lease.acquire(1, req);
-            assert_eq!(
-                granted, req,
-                "root lease must grant the full request while budget remains",
-            );
-            total += granted;
-            remaining -= granted;
-        }
-        assert_eq!(total, expected_total);
-        // Budget fully drained — any further acquire must return 0.
-        assert_eq!(lease.acquire(1, 1), 0);
-    }
-
-    #[test]
-    fn build_shared_cos_root_leases_reuses_existing_matching_lease_arc() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 100_000_000,
-                burst_bytes: 256 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![CoSQueueConfig {
-                    queue_id: 0,
-                    forwarding_class: "best-effort".into(),
-                    priority: 5,
-                    transmit_rate_bytes: 100_000_000,
-                    exact: false,
-                    surplus_weight: 1,
-                    buffer_bytes: 128 * 1024,
-                    dscp_rewrite: None,
-                }],
-            },
-        );
-        let active_shards_by_egress_ifindex = BTreeMap::from([(80, 1usize)]);
-
-        let existing = build_shared_cos_root_leases(&forwarding, &active_shards_by_egress_ifindex);
-        let reused = build_shared_cos_root_leases_reusing_existing(
-            &forwarding,
-            &active_shards_by_egress_ifindex,
-            &existing,
-        );
-
-        assert!(Arc::ptr_eq(
-            existing.get(&80).expect("existing lease"),
-            reused.get(&80).expect("reused lease")
-        ));
-    }
-
-    #[test]
-    fn build_shared_cos_queue_leases_reuses_existing_matching_lease_arc() {
-        let mut forwarding = ForwardingState::default();
-        forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 100_000_000,
-                burst_bytes: 256 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![CoSQueueConfig {
-                    queue_id: 4,
-                    forwarding_class: "iperf-b".into(),
-                    priority: 5,
-                    transmit_rate_bytes: 50_000_000,
-                    exact: true,
-                    surplus_weight: 1,
-                    buffer_bytes: 128 * 1024,
-                    dscp_rewrite: None,
-                }],
-            },
-        );
-        let active_shards_by_egress_ifindex = BTreeMap::from([(80, 2usize)]);
-
-        let existing = build_shared_cos_queue_leases_reusing_existing(
-            &forwarding,
-            &active_shards_by_egress_ifindex,
-            &BTreeMap::new(),
-        );
-        let reused = build_shared_cos_queue_leases_reusing_existing(
-            &forwarding,
-            &active_shards_by_egress_ifindex,
-            &existing,
-        );
-
-        assert!(Arc::ptr_eq(
-            existing.get(&(80, 4)).expect("existing queue lease"),
-            reused.get(&(80, 4)).expect("reused queue lease")
-        ));
-    }
-
-    #[test]
-    fn refresh_cos_owner_worker_map_from_binding_statuses_keeps_shared_arcs_when_unchanged() {
-        let mut coordinator = Coordinator::new();
-        coordinator.forwarding.cos.interfaces.insert(
-            80,
-            CoSInterfaceConfig {
-                shaping_rate_bytes: 100_000_000,
-                burst_bytes: 256 * 1024,
-                default_queue: 0,
-                dscp_classifier: String::new(),
-                ieee8021_classifier: String::new(),
-                dscp_queue_by_dscp: [u8::MAX; 64],
-                ieee8021_queue_by_pcp: [u8::MAX; 8],
-                queue_by_forwarding_class: FastMap::default(),
-                queues: vec![
-                    CoSQueueConfig {
-                        queue_id: 0,
-                        forwarding_class: "best-effort".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 50_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 128 * 1024,
-                        dscp_rewrite: None,
-                    },
-                    CoSQueueConfig {
-                        queue_id: 1,
-                        forwarding_class: "af11".into(),
-                        priority: 5,
-                        transmit_rate_bytes: 50_000_000,
-                        exact: false,
-                        surplus_weight: 1,
-                        buffer_bytes: 128 * 1024,
-                        dscp_rewrite: None,
-                    },
-                ],
-            },
-        );
-        coordinator.forwarding.egress.insert(
-            80,
-            EgressInterface {
-                bind_ifindex: 12,
-                vlan_id: 0,
-                mtu: 1500,
-                src_mac: [0; 6],
-                zone_id: TEST_WAN_ZONE_ID,
-                redundancy_group: 0,
-                primary_v4: None,
-                primary_v6: None,
-            },
-        );
-        let bindings = vec![BindingStatus {
-            worker_id: 7,
-            ifindex: 12,
-            ready: true,
-            ..Default::default()
-        }];
-
-        coordinator.refresh_cos_owner_worker_map_from_binding_statuses(&bindings);
-        let owners_before = coordinator.cos.owner_worker_by_queue.load_full();
-        let leases_before = coordinator.cos.root_leases.load_full();
-        let lease_before = leases_before.get(&80).expect("shared root lease").clone();
-        assert_eq!(lease_before.acquire(1, 2500), 2500);
-
-        coordinator.refresh_cos_owner_worker_map_from_binding_statuses(&bindings);
-        let owners_after = coordinator.cos.owner_worker_by_queue.load_full();
-        let leases_after = coordinator.cos.root_leases.load_full();
-
-        assert!(Arc::ptr_eq(&owners_before, &owners_after));
-        assert!(Arc::ptr_eq(
-            &lease_before,
-            leases_after.get(&80).expect("shared root lease")
-        ));
-    }
-
-    #[test]
-    fn unique_interface_owner_worker_id_returns_none_when_queues_split() {
-        let owner = 7u32;
-        let queues = vec![
-            crate::protocol::CoSQueueStatus {
-                queue_id: 0,
-                owner_worker_id: Some(2),
-                ..Default::default()
-            },
-            crate::protocol::CoSQueueStatus {
-                queue_id: 1,
-                owner_worker_id: Some(owner),
-                ..Default::default()
-            },
-        ];
-
-        assert_eq!(unique_interface_owner_worker_id(&queues), None);
-    }
-
-    #[test]
-    fn aggregate_cos_statuses_sums_drop_counters_across_worker_snapshots() {
-        // #710 regression pin. This is the EXACT code path where the
-        // live bug landed: `Coordinator::cos_statuses` re-aggregates
-        // per-worker snapshots, and before this PR that re-aggregation
-        // silently dropped every new drop-counter field. The unit test
-        // gate must be at this layer, not just the worker layer.
-        use crate::protocol::{CoSInterfaceStatus, CoSQueueStatus};
-
-        let worker_a = vec![CoSInterfaceStatus {
-            ifindex: 80,
-            interface_name: "reth0.80".into(),
-            shaping_rate_bytes: 1_250_000_000,
-            burst_bytes: 256 * 1024,
-            worker_instances: 1,
-            queues: vec![CoSQueueStatus {
-                queue_id: 4,
-                worker_instances: 1,
-                admission_flow_share_drops: 3,
-                admission_buffer_drops: 5,
-                admission_ecn_marked: 37,
-                root_token_starvation_parks: 7,
-                queue_token_starvation_parks: 11,
-                tx_ring_full_submit_stalls: 13,
-                ..Default::default()
-            }],
-            ..Default::default()
-        }];
-        let worker_b = vec![CoSInterfaceStatus {
-            ifindex: 80,
-            interface_name: "reth0.80".into(),
-            shaping_rate_bytes: 1_250_000_000,
-            burst_bytes: 256 * 1024,
-            worker_instances: 1,
-            queues: vec![CoSQueueStatus {
-                queue_id: 4,
-                worker_instances: 1,
-                admission_flow_share_drops: 17,
-                admission_buffer_drops: 19,
-                admission_ecn_marked: 41,
-                root_token_starvation_parks: 23,
-                queue_token_starvation_parks: 29,
-                tx_ring_full_submit_stalls: 31,
-                ..Default::default()
-            }],
-            ..Default::default()
-        }];
-        let owner_by_queue = BTreeMap::from([((80, 4u8), 3u32)]);
-        let aggregated =
-            aggregate_cos_statuses_across_workers(&[worker_a, worker_b], &owner_by_queue);
-
-        assert_eq!(aggregated.len(), 1);
-        let iface = &aggregated[0];
-        assert_eq!(iface.ifindex, 80);
-        assert_eq!(iface.queues.len(), 1);
-        let q = &iface.queues[0];
-        assert_eq!(q.queue_id, 4);
-        assert_eq!(q.owner_worker_id, Some(3));
-        // Each counter is non-coprime-prime on both sides to catch
-        // accidental re-attribution between counters.
-        assert_eq!(q.admission_flow_share_drops, 3 + 17);
-        assert_eq!(q.admission_buffer_drops, 5 + 19);
-        assert_eq!(q.admission_ecn_marked, 37 + 41);
-        assert_eq!(q.root_token_starvation_parks, 7 + 23);
-        assert_eq!(q.queue_token_starvation_parks, 11 + 29);
-        assert_eq!(q.tx_ring_full_submit_stalls, 13 + 31);
-    }
-
-    #[test]
-    fn aggregate_cos_statuses_sums_owner_profile_across_workers_coherently() {
-        use crate::protocol::{CoSInterfaceStatus, CoSQueueStatus};
-
-        let worker_a = vec![CoSInterfaceStatus {
-            ifindex: 80,
-            interface_name: "reth0.80".into(),
-            worker_instances: 1,
-            queues: vec![CoSQueueStatus {
-                queue_id: 4,
-                worker_instances: 1,
-                exact: true,
-                drain_latency_hist: {
-                    let mut v = vec![0; super::super::umem::DRAIN_HIST_BUCKETS];
-                    v[0] = 5;
-                    v
-                },
-                redirect_acquire_hist: {
-                    let mut v = vec![0; super::super::umem::DRAIN_HIST_BUCKETS];
-                    v[1] = 3;
-                    v
-                },
-                drain_invocations: 5,
-                drain_noop_invocations: 1,
-                owner_pps: 100,
-                peer_pps: 40,
-                ..Default::default()
-            }],
-            ..Default::default()
-        }];
-        let worker_b = vec![CoSInterfaceStatus {
-            ifindex: 80,
-            interface_name: "reth0.80".into(),
-            worker_instances: 1,
-            queues: vec![CoSQueueStatus {
-                queue_id: 4,
-                worker_instances: 1,
-                exact: true,
-                drain_latency_hist: {
-                    let mut v = vec![0; super::super::umem::DRAIN_HIST_BUCKETS];
-                    v[7] = 11;
-                    v
-                },
-                redirect_acquire_hist: {
-                    let mut v = vec![0; super::super::umem::DRAIN_HIST_BUCKETS];
-                    v[2] = 13;
-                    v
-                },
-                drain_invocations: 11,
-                drain_noop_invocations: 2,
-                owner_pps: 200,
-                peer_pps: 50,
-                ..Default::default()
-            }],
-            ..Default::default()
-        }];
-
-        let owner_by_queue = BTreeMap::from([((80, 4u8), 3u32)]);
-        let aggregated =
-            aggregate_cos_statuses_across_workers(&[worker_a, worker_b], &owner_by_queue);
-
-        let q = &aggregated[0].queues[0];
-        assert_eq!(q.drain_latency_hist[0], 5);
-        assert_eq!(q.drain_latency_hist[7], 11);
-        assert_eq!(q.redirect_acquire_hist[1], 3);
-        assert_eq!(q.redirect_acquire_hist[2], 13);
-        assert_eq!(q.drain_invocations, 16);
-        assert_eq!(q.drain_noop_invocations, 3);
-        assert_eq!(q.owner_pps, 300);
-        assert_eq!(q.peer_pps, 90);
-        assert_eq!(
-            q.drain_latency_hist.iter().copied().sum::<u64>(),
-            q.drain_invocations,
-            "cross-worker aggregation must preserve hist == invocation invariant",
-        );
-    }
-
-    #[test]
-    fn cos_no_owner_binding_drops_total_sums_across_every_live_state() {
-        // #710: the per-binding `no_owner_binding_drops` atomic is the
-        // mechanical accumulator; the operator-facing surface is
-        // `Coordinator::cos_no_owner_binding_drops_total`, which must
-        // sum across every `BindingLiveState`. Without this test, a
-        // refactor that reads only `bindings.first()` or only one
-        // worker's bindings could silently undercount.
-        let a = std::sync::Arc::new(BindingLiveState::new());
-        let b = std::sync::Arc::new(BindingLiveState::new());
-        let c = std::sync::Arc::new(BindingLiveState::new());
-        a.no_owner_binding_drops
-            .store(3, std::sync::atomic::Ordering::Relaxed);
-        b.no_owner_binding_drops
-            .store(5, std::sync::atomic::Ordering::Relaxed);
-        c.no_owner_binding_drops
-            .store(7, std::sync::atomic::Ordering::Relaxed);
-
-        let total: u64 = [a, b, c]
-            .iter()
-            .map(|live| {
-                live.no_owner_binding_drops
-                    .load(std::sync::atomic::Ordering::Relaxed)
-            })
-            .sum();
-        assert_eq!(total, 15);
-    }
-
-    #[test]
-    fn ring_pressure_counters_round_trip_through_snapshot() {
-        // #802: verify that the new ring-pressure atomics on
-        // BindingLiveState are surfaced via `snapshot()`. Without this
-        // pin, a refactor that drops the new fields from `snapshot()`
-        // would silently zero the operator-facing counters.
-        use std::sync::atomic::Ordering;
-        let live = BindingLiveState::new();
-        live.dbg_tx_ring_full.store(11, Ordering::Relaxed);
-        live.dbg_sendto_enobufs.store(13, Ordering::Relaxed);
-        // #804: two distinct counters — bound-pending FIFO overflow
-        // (17) and CoS queue admission overflow (41). Non-coprime-prime
-        // per field so an accidental swap across the two is caught.
-        live.dbg_bound_pending_overflow.store(17, Ordering::Relaxed);
-        live.dbg_cos_queue_overflow.store(41, Ordering::Relaxed);
-        live.rx_fill_ring_empty_descs.store(19, Ordering::Relaxed);
-        live.debug_outstanding_tx.store(23, Ordering::Relaxed);
-        let snap = live.snapshot();
-        assert_eq!(snap.dbg_tx_ring_full, 11);
-        assert_eq!(snap.dbg_sendto_enobufs, 13);
-        assert_eq!(snap.dbg_bound_pending_overflow, 17);
-        assert_eq!(snap.dbg_cos_queue_overflow, 41);
-        assert_eq!(snap.rx_fill_ring_empty_descs, 19);
-        assert_eq!(snap.debug_outstanding_tx, 23);
-    }
-
-    // -------------------------------------------------------------
-    // #925 Phase 1: worker supervisor catch_unwind tests.
-    // -------------------------------------------------------------
-
-    /// Helper: extract the message from a caught panic payload using
-    /// the same renderer the supervisor uses.
-    fn caught_message<F: FnOnce() + std::panic::UnwindSafe>(f: F) -> String {
-        let r = std::panic::catch_unwind(f);
-        let payload = r.unwrap_err();
-        super::panic_payload_message(&payload)
-    }
-
-    #[test]
-    fn panic_payload_message_renders_str_panic() {
-        assert_eq!(caught_message(|| panic!("hello world")), "hello world");
-    }
-
-    #[test]
-    fn panic_payload_message_renders_string_panic() {
-        let s = String::from("owned message");
-        assert_eq!(caught_message(move || panic!("{}", s)), "owned message");
-    }
-
-    #[test]
-    fn panic_payload_message_falls_back_for_non_string() {
-        // panic_any unwinds with a non-string payload (i32 here).
-        let msg = caught_message(|| std::panic::panic_any(42_i32));
-        assert_eq!(msg, "non-string panic payload");
-    }
-
-    /// Integration test against the same `spawn_supervised_worker`
-    /// production uses (the spawn-closure body is the only thing we
-    /// substitute — the supervisor wrapper is the real one).
-    #[test]
-    fn spawn_supervised_worker_catches_string_panic_and_marks_dead() {
-        use std::sync::atomic::Ordering;
-        let atomics = Arc::new(super::super::worker_runtime::WorkerRuntimeAtomics::new());
-        let slot = Arc::new(Mutex::new(None::<String>));
-        let join = super::spawn_supervised_worker(
-            7,
-            atomics.clone(),
-            slot.clone(),
-            || panic!("intentional test panic"),
-        )
-        .expect("spawn_supervised_worker");
-        // The supervisor must NOT propagate the panic to the joiner.
-        join.join().expect("supervisor must catch worker panic");
-        assert!(atomics.dead.load(Ordering::Relaxed));
-        let msg = slot
-            .lock()
-            .expect("panic slot lock")
-            .clone()
-            .expect("panic message published");
-        assert_eq!(msg, "intentional test panic");
-    }
-}
+mod tests;
